Chaos : fix second compilation issue
#rb trivial
[FYI] jaco.vandyk
#jira none

#ROBOMERGE-AUTHOR: cedric.caillaud
#ROBOMERGE-SOURCE: CL 18110229 in //UE4/Release-4.27-Chaos/... via CL 18110281 via CL 18110335
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v889-18060218)

[CL 18110440 by cedric caillaud in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/Experimental/Chaos/Public/Chaos/AABBTreeDirtyGridUtils.h
@@ -276,8 +276,8 @@ namespace Chaos
 		bool DxTooSmall = AbsDx <= SMALL_NUMBER;
 		bool DyTooSmall = AbsDy <= SMALL_NUMBER;
 
-		int DeltaCelIndexX;
-		int DeltaCelIndexY;
+		int64 DeltaCelIndexX;
+		int64 DeltaCelIndexY;
 		FReal DtDy = 0; // DeltaTime over DeltaX
 		FReal DtDx = 0;
 
@@ -343,7 +343,7 @@ namespace Chaos
 				}
 			}
 
-			for (int CurrentFillCellIndexY = CurrentCellIndexY1; CurrentFillCellIndexY * DeltaCelIndexY <= CurrentCellIndexY0 * DeltaCelIndexY; CurrentFillCellIndexY += DeltaCelIndexY)
+			for (int64 CurrentFillCellIndexY = CurrentCellIndexY1; CurrentFillCellIndexY * DeltaCelIndexY <= CurrentCellIndexY0 * DeltaCelIndexY; CurrentFillCellIndexY += DeltaCelIndexY)
 			{
 				InFunction((FReal)CurrentCellIndexX0 * DirtyElementGridCellSize, (FReal)CurrentFillCellIndexY * DirtyElementGridCellSize);
 			}
